django consuming rabbitMQ with celery
