+++ 11_servidor_ose_1/dockerfile
@@ -11,7 +11,9 @@ WORKDIR /app/ose_core/
 # Install dependencies
 COPY requirements.txt /app/
 RUN pip install --no-cache-dir -r ../requirements.txt
-#RUN pip install --no-cache-dir -r ../requirements.txt
+
+# Install Gunicorn
+RUN pip install gunicorn
 
 # Copy the Django project code to the container
 COPY . /app/
@@ -19,5 +21,5 @@ COPY . /app/
 # Expose the port that Django runs on (default: 8000)
 EXPOSE 8000
 
-# Run the Django development server
-CMD ["python", "manage.py", "runserver", "0.0.0.0:8000"]
+# Run the Celery worker and the Django application using Gunicorn
+CMD celery -A ose_core worker --loglevel=info & gunicorn ose_core.wsgi:application --bind 0.0.0.0:8000 --log-level debug
