+++ 11_servidor_ose_1/ose_core/ose_core/__init__.py
@@ -1,0 +1,3 @@
+from .celery import app as celery_app
+
+__all__ = ['celery_app']
