+++ 11_servidor_ose_1/ose_core/ose_core/celery.py
@@ -1,0 +1,14 @@
+import os
+from celery import Celery
+
+# Set the default Django settings module for Celery
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'ose_core.settings')
+
+# Create the Celery app
+app = Celery('ose_core')
+
+# Configure Celery
+app.config_from_object('django.conf:settings', namespace='CELERY')
+
+# Autodiscover tasks from all registered Django apps
+app.autodiscover_tasks()
