+++ 11_servidor_ose_1/requirements.txt
@@ -2,3 +2,4 @@ pika==1.3.2
 asgiref==3.6.0
 Django==4.2.1
 sqlparse==0.4.4
+celery==5.1.2
