+++ docker-compose.yml
@@ -38,12 +38,15 @@ services:
 
   11_servidor_ose_1:
     build: ./11_servidor_ose_1
+    ports:
+      - "8000:8000"
     environment:
       - ALERTS_QUEUE_PRIMARY=alerts-queue-1
       - ALERTS_QUEUE_SECONDARY=alerts-queue-2
       - QUEUE_NAME=test_queue
       - PYTHONUNBUFFERED=1
     depends_on:
-      alerts-queue-1:
-        condition: service_healthy
+    - alerts-queue-1
+    - alerts-queue-2
+
